Make filtering by hook type use element wise check instead of substring match
I just checked, which `tex` hooks are available on the website. It seems that filtering on `tex` also shows all the results for `text`. The list is almost identical between the two types.

If I understand the code correctly `hookTypes` is something like `shell, non-executable`. By splitting it on `', '` the `includes` check will work element wise instead of performing a substring match.